pyproject.toml
[project]
name = "hyper-lrp"
version = "0.1.0"
description = "Reusable LRP tooling for vision-language models, with HyperView integration helpers"
readme = "README.md"
authors = [
    { name = "hyper3labs" }
]
license = "MIT"
requires-python = ">=3.11.11"
dependencies = [
    "lxt>=2.1,<3",
    "numpy>=1.26,<3",
    "Pillow>=10,<13",
    "torch>=2.6,<3",
    "transformers==4.52.4",
    "zennit>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
hyper-lrp = "hyper_lrp.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["vendor", "artifacts", "context", ".venv"]

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"
